Update UserQPTemplateDocument schema and connection string
- Changed `UserQPTemplateDocumentId` type from `INT` to `BIGINT`.
- Removed database save changes line in `QpTemplateService.cs`.
- Updated connection string in `appsettings.Development.json` for local SQL Server.

+++ Database/SECE.Database/Tables/UserQPTemplateDocument.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[UserQPTemplateDocument]
 (
-	[UserQPTemplateDocumentId] INT NOT NULL PRIMARY KEY IDENTITY(1,1),
+	[UserQPTemplateDocumentId] BIGINT NOT NULL PRIMARY KEY IDENTITY(1,1),
 	[UserQPTemplateId] BIGINT NOT NULL FOREIGN KEY (UserQPTemplateId) REFERENCES [UserQPTemplate](UserQPTemplateId),
 	[QPDocumentTypeId] BIGINT NOT NULL,
 	[DocumentId] BIGINT NOT NULL,
